Tidied helper functions
