+++ app/controllers/helpers/storage.py
@@ -2,7 +2,6 @@ from app import app
 from boto.s3.connection import S3Connection
 from boto.s3.key import Key
 import json
-import re
 
 
 class IneffableStorage(object):
